requirements.txt
langchain
openai
streamlit
jupyter
tiktoken
faiss-cpu
python-dotenv
ipython